+++ src/modules.js
@@ -97,6 +97,7 @@ global.LibraryManager = {
       libraries.push('library_wasmfs.js');
       libraries.push('library_wasmfs_js_file.js');
       libraries.push('library_wasmfs_fetch.js');
+      libraries.push('library_wasmfs_node.js');
     }
 
     // Additional JS libraries (without AUTO_JS_LIBRARIES, link to these explicitly via -lxxx.js)