+++ src/parseTools.js
@@ -118,6 +118,14 @@ function preprocess(text, filenameHint) {
             } else if (first === '#endif') {
               assert(showStack.length > 0);
               showStack.pop();
+            } else if (first === '#warning') {
+              if (showCurrentLine()) {
+                printErr(`${filenameHint}:${i + 1}: #warning ${trimmed.substring(trimmed.indexOf(' ')).trim()}`);
+              }
+            } else if (first === '#error') {
+              if (showCurrentLine()) {
+                error(`${filenameHint}:${i + 1}: #error ${trimmed.substring(trimmed.indexOf(' ')).trim()}`);
+              }
             } else {
               throw new Error(`Unknown preprocessor directive on line ${i}: ``${line}```);
             }