+++ src/preamble.js
@@ -137,6 +137,10 @@ function ccall(ident, returnType, argTypes, args, opts) {
       }
     }
   }
+#if ASYNCIFY
+  // Data for a previous async operation that was in flight before us.
+  var previousAsync = Asyncify.currData;
+#endif
   var ret = func.apply(null, cArgs);
   function onDone(ret) {
 #if ASYNCIFY
@@ -146,11 +150,22 @@ function ccall(ident, returnType, argTypes, args, opts) {
     return convertReturnValue(ret);
   }
 #if ASYNCIFY
+  // Keep the runtime alive through all calls. Note that this call might not be
+  // async, but for simplicity we push and pop in all calls.
   runtimeKeepalivePush();
   var asyncMode = opts && opts.async;
-  // Check if we started an async operation just now.
-  if (Asyncify.currData) {
-    // If so, the WASM function ran asynchronous and unwound its stack.
+  if (Asyncify.currData != previousAsync) {
+#if ASSERTIONS
+    // A change in async operation happened. If there was already an async
+    // operation in flight before us, that is an error: we should not start
+    // another async operation while one is active, and we should not stop one
+    // either. The only valid combination is to have no change in the async
+    // data (so we either had one in flight and left it alone, or we didn't have
+    // one), or to have nothing in flight and to start one.
+    assert(!(previousAsync && Asyncify.currData), 'We cannot start an async operation when one is already flight');
+    assert(!(previousAsync && !Asyncify.currData), 'We cannot stop an async operation in flight');
+#endif
+    // This is a new async operation. The wasm is paused and has unwound its stack.
     // We need to return a Promise that resolves the return value
     // once the stack is rewound and execution finishes.
 #if ASSERTIONS
@@ -616,7 +631,7 @@ function abort(what) {
 
 #include "memoryprofiler.js"
 
-#if ASSERTIONS && !('$FS' in addedLibraryItems) && !ASMFS && !WASMFS
+#if ASSERTIONS && !('$FS' in addedLibraryItems) && !WASMFS
 // show errors on likely calls to FS when it was not included
 var FS = {
   error: function() {