+++ ChangeLog.md
@@ -18,8 +18,14 @@ to browse the changes between the tags.
 
 See docs/process.md for more on how version tagging works.
 
-3.1.6
+3.1.7
 -----
+- Remove unmaintained ASMFS filesystem backend and associated `-sASMFS`
+  settings.  The new wasmfs filesystem is far enough along that it seems clear
+  that ASMFS will not need to be revived.
+
+3.1.6 - 02/24/2022
+------------------
 - Remove support for deprecated `EMMAKEN_COMPILER`, `EMMAKEN_CFLAGS`, and
   `EMMAKEN_NO_SDK`  environment variables.  These are all legacy and redundant
   in the face of other settings/flags: